update(blog/ assets/ download/ index/): Update to include information about 2022.05.28 release
Update to include information about the latest release. The download page has been updated to prepare for the release information and links. Blog page has been updated to include a blog post detailing the changes to the disk image. A new image is used due to the change in wallpaper for the homescreen. The home screen has new information about the minimal disk image.